Move SmartBlock out of the MrIceBlock file.
