+++ src/badguy/mriceblock.cpp
@@ -297,11 +297,4 @@ MrIceBlock::ignite() {
   BadGuy::ignite();
 }
 
-SmartBlock::SmartBlock(const ReaderMapping& reader) :
-  MrIceBlock(reader)
-{
-  max_drop_height = 16;
-  sprite = SpriteManager::current()->create("images/creatures/mr_iceblock/smart_block/smart_block.sprite");
-}
-
 /* EOF */
