+++ src/supertux/object_factory.cpp
@@ -28,7 +28,6 @@
 #include "badguy/angrystone.hpp"
 #include "badguy/badguy.hpp"
 #include "badguy/bomb.hpp"
-#include "badguy/skydive.hpp"
 #include "badguy/bouncing_snowball.hpp"
 #include "badguy/captainsnowball.hpp"
 #include "badguy/crystallo.hpp"
@@ -59,7 +58,9 @@
 #include "badguy/root.hpp"
 #include "badguy/short_fuse.hpp"
 #include "badguy/skullyhop.hpp"
+#include "badguy/skydive.hpp"
 #include "badguy/smartball.hpp"
+#include "badguy/smartblock.hpp"
 #include "badguy/snail.hpp"
 #include "badguy/snowball.hpp"
 #include "badguy/snowman.hpp"
@@ -94,7 +95,6 @@
 #include "object/bullet.hpp"
 #include "object/camera.hpp"
 #include "object/candle.hpp"
-#include "object/torch.hpp"
 #include "object/cloud_particle_system.hpp"
 #include "object/coin.hpp"
 #include "object/coin_explode.hpp"
@@ -102,8 +102,8 @@
 #include "object/decal.hpp"
 #include "object/display_effect.hpp"
 #include "object/electrifier.hpp"
-#include "object/endsequence_fireworks.hpp"
 #include "object/endsequence.hpp"
+#include "object/endsequence_fireworks.hpp"
 #include "object/endsequence_walkleft.hpp"
 #include "object/endsequence_walkright.hpp"
 #include "object/explosion.hpp"
@@ -155,6 +155,7 @@
 #include "object/text_object.hpp"
 #include "object/thunderstorm.hpp"
 #include "object/tilemap.hpp"
+#include "object/torch.hpp"
 #include "object/trampoline.hpp"
 #include "object/unstable_tile.hpp"
 #include "object/weak_block.hpp"
